fix: int -> Long, add root gitignore

+++ back/tripDuo/src/main/java/com/example/tripDuo/controller/UserController.java
@@ -60,7 +60,7 @@ public class UserController {
 
     
     @DeleteMapping("/{id}")
-    public ResponseEntity<String> deleteUser(@PathVariable int id) {
+    public ResponseEntity<String> deleteUser(@PathVariable Long id) {
         service.deleteUser(id);
     	return ResponseEntity.ok(id + " user is deleted");
     }
